+++ packages/solid/src/components/collection/index.tsx
@@ -14,4 +14,4 @@ export {
   type TreeCollectionOptions,
   type TreeNode,
 } from './tree-collection'
-export { useListCollection, type UseListCollectionProps } from './use-list-collection'
+export { useListCollection, type UseListCollectionProps, type UseListCollectionReturn } from './use-list-collection'