+++ packages/svelte/CHANGELOG.md
@@ -6,6 +6,10 @@ description: All notable changes will be documented in this file.
 
 ## [Unreleased]
 
+### Fixed
+
+- **Collection**: Export `UseListCollectionReturn` type
+
 ## [5.1.1] - 2025-07-05
 
 ### Fixed